requirements.txt
aioredis==1.3.1
amqp==5.0.1
aniso8601==8.0.0
asgiref==3.3.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
cffi==1.13.1
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-repl==0.1.6
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
daphne==2.5.0
DBUtils==1.3
decorator==4.4.2
Django==2.2.5
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-js-asset==1.2.2
django-qiniu-storage==2.3.1
django-redis==4.12.1
django-simpleui==2020.9.26
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Flask==1.0.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.7
gevent==1.4.0
greenlet==0.4.15
hiredis==1.1.0
html5lib==1.1
hyperlink==19.0.0
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.10.0
Js2Py==0.70
jsonpath==0.82
kombu==5.0.2
lxml==4.5.2
MarkupSafe==1.1.1
MouseInfo==0.1.3
msgpack==1.0.0
mysqlclient==1.4.4
numpy==1.19.4
opencv-python==4.4.0.46
optionaldict==0.1.1
parsel==1.6.0
pdfdocument==4.0.0
pdfminer3k==1.3.4
Pillow==6.2.1
ply==3.11
prompt-toolkit==3.0.7
Protego==0.1.16
psutil==5.7.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycparser==2.19
pycryptodome==3.9.8
PyDispatcher==2.0.5
PyGetWindow==0.0.8
PyHamcrest==2.0.2
pyjsparser==2.7.1
PyJWT==1.7.1
PyMsgBox==1.0.8
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
pytesseract==0.3.4
python-dateutil==2.8.1
python-weixin==0.5.0
PyTweening==1.0.3
pytz==2019.3
qiniu==7.3.0
queuelib==1.5.0
redis==3.5.3
reportlab==3.5.46
requests==2.24.0
retry==0.9.2
retrying==1.3.3
ronglian-sms-sdk==1.0
Scrapy==2.3.0
selenium==3.141.0
service-identity==18.1.0
simplejson==3.17.2
six==1.13.0
soupsieve==2.0.1
sqlparse==0.3.0
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
urllib3==1.25.9
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
wechatpy==1.8.14
Werkzeug==0.16.0
wfastcgi==3.0.0
xmltodict==0.12.0
zipp==3.3.0
zope.interface==5.1.0
